backend/requirements.txt
# Python dependencies
fastapi==0.95.2
pydantic==1.10.7
uvicorn==0.23.1
beautifulsoup4==4.12.2
requests==2.31.0
google-generativeai==0.3.2
python-dotenv==1.0.0
pytest
pytest-asyncio
httpx